+++ featurebyte/routes/feature_list/api.py
@@ -9,7 +9,7 @@ from typing import Any, Dict, Optional, Union, cast
 import json
 from http import HTTPStatus
 
-from fastapi import APIRouter, File, Form, Request, UploadFile
+from fastapi import APIRouter, File, Form, Query, Request, UploadFile
 from fastapi.responses import ORJSONResponse, StreamingResponse
 
 from featurebyte.models.base import PydanticObjectId
@@ -239,3 +239,21 @@ async def get_online_features(
         get_credential=request.state.get_credential,
     )
     return cast(OnlineFeaturesResponseModel, result)
+
+
+@router.get("/{feature_list_id}/feature_job_logs", response_model=Dict[str, Any])
+async def get_feature_job_logs(
+    request: Request,
+    feature_list_id: PydanticObjectId,
+    hour_limit: int = Query(default=24, gt=0, le=2400),
+) -> Dict[str, Any]:
+    """
+    Retrieve feature job status
+    """
+    controller = request.state.app_container.feature_list_controller
+    result = await controller.get_feature_job_logs(
+        feature_list_id=feature_list_id,
+        hour_limit=hour_limit,
+        get_credential=request.state.get_credential,
+    )
+    return cast(Dict[str, Any], result)